# Blog Project（Frontend）/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BinSpace API
  version: 2.0.0-draft.1
  description: |
    BinSpace 当前 42 个接口基线与下一阶段功能 2～8 契约。
    认证使用 Spring Security、Bearer JWT 和 Redis Token Session。
    新增接口在后端实现前仅代表约定，不代表当前环境已经可调用。
servers:
  - url: http://localhost:8080
    description: Local Spring Boot
...
paths:
...
  /api/owner/site/basic:
    put:
      summary: 更新空间基本信息
      description: 更新站点英文名、中文名与说明，使用 site_config 共享 version。
      tags: [Owner Site]
      operationId: updateSiteBasic
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSiteBasicDTO' }
      responses:
        '200':
          description: 更新后的 Basic 与最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SiteBasicSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/site/hero:
    put:
      summary: 更新 Hero
      description: 保存 Hero 文案及 desktop/mobile video/poster URL，使用 site_config 共享 version。
      tags: [Owner Site]
      operationId: updateHero
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateHeroDTO' }
      responses:
        '200':
          description: 更新后的 Hero 与最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HeroSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/site/announcement:
    put:
      summary: 更新空间公告
      description: enabled=true 时 content 必须非空；使用 site_config 共享 version。
      tags: [Owner Site]
      operationId: updateAnnouncement
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAnnouncementDTO' }
      responses:
        '200':
          description: 更新后的公告或 null 及最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnouncementSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/site/music:
    put:
      summary: 更新空间音乐
      description: audioUrl 为空表示关闭音乐；使用 site_config 共享 version。
      tags: [Owner Site]
      operationId: updateMusic
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMusicDTO' }
      responses:
        '200':
          description: 更新后的音乐或 null 及最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MusicSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/site/appearance:
    put:
      summary: 更新默认空间外观
      description: 保存全站默认外观；Visitor 的页面内临时偏好不调用此接口。使用共享 version。
      tags: [Owner Site]
      operationId: updateAppearance
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAppearanceDTO' }
      responses:
        '200':
          description: 更新后的外观与最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppearanceSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/site/page-media:
    put:
      summary: 更新页面媒体配置
      description: 保存 Moments、Guestbook、Tech、Archive、About 背景配置，使用共享 version。
      tags: [Owner Site]
      operationId: updatePageMedia
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePageMediaDTO' }
      responses:
        '200':
          description: 更新后的页面媒体与最新共享版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageMediaSectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
...
  /api/tags:
    get:
      summary: 获取公开标签
      description: 只返回 active 标签；used=true 时仅返回被公开 TECH 使用的标签；无请求体。
      tags: [Tags]
      operationId: listPublicTags
      parameters:
        - { $ref: '#/components/parameters/UsedQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页标签
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/owner/tags:
    get:
      summary: 获取 Owner 标签列表
      description: 支持 active/disabled/all 管理视图；无请求体。
      tags: [Tags]
      operationId: listOwnerTags
      parameters:
        - { $ref: '#/components/parameters/ActiveStatusQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页标签管理数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagAdminPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    post:
      summary: 创建标签
      description: slug 为空时服务端生成；name/slug 必须唯一。
      tags: [Tags]
      operationId: createTag
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTagDTO' }
      responses:
        '201':
          description: 创建后的标签
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagAdminResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/tags/{id}:
    put:
      summary: 更新标签
      description: 当前 tag 表没有 version，不虚构乐观锁字段。
      tags: [Tags]
      operationId: updateTag
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTagDTO' }
      responses:
        '200':
          description: 更新后的标签
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagAdminResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    delete:
      summary: 停用标签
      description: 逻辑删除，不删除 post_tag 历史关联；无请求体。
      tags: [Tags]
      operationId: deleteTag
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '204': { description: 停用成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录/注册接口返回的 JWT Access Token。Visitor 与 Owner 受保护请求均使用 Authorization: Bearer <token>。"
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
...
    UsedQuery:
      name: used
      in: query
      required: false
      schema: { type: boolean }
      description: true 时仅返回被公开 TECH 使用的项目。
    ActiveStatusQuery:
      name: status
      in: query
      required: false
      schema: { type: string, enum: [ACTIVE, DISABLED, ALL], default: ALL }
...
  responses:
    BadRequest:
      description: 请求参数或 Bean Validation 失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: BAD_REQUEST, message: 请求参数不正确, data: { title: 标题不能为空 }, traceId: 01JEXAMPLE }
    Unauthorized:
      description: 未携带 Token、JWT 无效/过期，或 Redis Token Session 已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: AUTH_REQUIRED, message: 请先登录, data: null, traceId: 01JEXAMPLE }
...
    Forbidden:
      description: 当前 JWT 身份缺少接口要求的角色或权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: FORBIDDEN, message: 无权执行此操作, data: null, traceId: 01JEXAMPLE }
    NotFound:
      description: 资源不存在、已删除或不可见
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_NOT_FOUND, message: 资源不存在, data: null, traceId: 01JEXAMPLE }
    VersionConflict:
      description: 乐观锁版本冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: VERSION_CONFLICT, message: 数据已被修改, data: null, traceId: 01JEXAMPLE }
    ResourceConflict:
      description: 唯一键、重复回复或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_CONFLICT, message: 资源状态冲突, data: null, traceId: 01JEXAMPLE }
...
    InternalError:
      description: 未处理的服务端错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: INTERNAL_ERROR, message: 服务暂时不可用, data: null, traceId: 01JEXAMPLE }
...
  schemas:
    NullOnly:
      type: object
      nullable: true
      enum: [null]
      description: OpenAPI 3.0 中供 anyOf 复用的 null 分支。
    BigIntId:
      type: string
      pattern: '^[0-9]+$'
      description: MySQL BIGINT 的十进制字符串，禁止在前端转换为 Number。
      example: '1909508401234567168'
    ApiSuccessBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string, enum: [OK], example: OK }
        message: { type: string, example: success }
        traceId: { type: string, example: 01JEXAMPLE }
    ErrorResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, AUTH_REQUIRED, AUTH_FAILED, FORBIDDEN, RESOURCE_NOT_FOUND, VERSION_CONFLICT, RESOURCE_CONFLICT, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, RATE_LIMITED, UPSTREAM_ERROR, INTERNAL_ERROR]
        message: { type: string }
        data:
          type: object
          nullable: true
          additionalProperties: true
        traceId: { type: string }
...
    SiteBasicVO:
      type: object
      required: [name, chineseName, description]
      properties:
        name: { type: string, example: BinSpace }
        chineseName: { type: string, example: 玢的空间 }
        description: { type: string, nullable: true, example: 记录代码，也记录生活。 }
...
    HeroVO:
      type: object
      required: [eyebrow, heroTitle, heroSubtitle, desktopVideo, mobileVideo, poster]
      properties:
        eyebrow: { type: string, nullable: true, example: BINSPACE · PERSONAL DAYBOOK }
        heroTitle: { type: string, example: 在暖光里，留下代码与生活。 }
        heroSubtitle: { type: string, nullable: true, example: 这里收录认真写下的技术文章，也收藏日常时刻。 }
        desktopVideo: { type: string, nullable: true, example: /media/binspace-hero.mp4 }
        mobileVideo: { type: string, nullable: true, example: /media/binspace-hero-mobile.mp4 }
        poster: { type: string, nullable: true, example: /media/binspace-hero.jpg }
    AnnouncementVO:
      type: object
      required: [content, enabled]
      properties:
        content: { type: string, example: 欢迎来到玢的空间。 }
        enabled: { type: boolean, example: true }
    MusicVO:
      type: object
      required: [title, artist, audioUrl, cover]
      properties:
        title: { type: string, example: 空间背景音乐 }
        artist: { type: string, nullable: true, example: Artist }
        audioUrl: { type: string, example: /media/music/theme.mp3 }
        cover: { type: string, nullable: true, example: /media/music/theme.jpg }
    WallpaperVO:
      type: object
      required: [desktop, mobile]
      properties:
        desktop: { type: string, nullable: true, example: /media/pages/home-beyond-mountain.jpg }
        mobile: { type: string, nullable: true, example: /media/pages/home-beyond-mountain-mobile.jpg }
    AppearanceVO:
      type: object
      required: [layoutMode, backgroundMode, surfaceOpacity, backdropShade, wallpaper, allowVisitorControls]
      properties:
        layoutMode: { type: string, enum: [standard, immersive], example: standard }
        backgroundMode: { type: string, enum: [video, wallpaper], example: wallpaper }
        surfaceOpacity: { type: number, format: double, minimum: 0, maximum: 1, example: 0.54 }
        backdropShade: { type: number, format: double, minimum: 0, maximum: 1, example: 0.3 }
        wallpaper: { $ref: '#/components/schemas/WallpaperVO' }
        allowVisitorControls: { type: boolean, example: true }
    PageMediaItemVO:
      type: object
      required: [video, poster, overlay, motion, effect]
      properties:
        video: { type: string, nullable: true, example: /media/pages/archive-horizon-sky.mp4 }
        poster: { type: string, nullable: true, example: /media/pages/archive-horizon-sky.jpg }
        overlay: { type: string, nullable: true, example: null }
        motion: { type: string, nullable: true, example: drift }
        effect: { type: string, nullable: true, example: petals }
    PageMediaVO:
      type: object
      required: [moments, guestbook, tech, archive, about]
      properties:
        moments: { $ref: '#/components/schemas/PageMediaItemVO' }
        guestbook: { $ref: '#/components/schemas/PageMediaItemVO' }
        tech: { $ref: '#/components/schemas/PageMediaItemVO' }
        archive: { $ref: '#/components/schemas/PageMediaItemVO' }
        about: { $ref: '#/components/schemas/PageMediaItemVO' }
...
    UpdateSiteBasicDTO:
      type: object
      required: [name, chineseName, description, version]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        chineseName: { type: string, minLength: 1, maxLength: 64 }
        description: { type: string, nullable: true, maxLength: 255 }
        version: { type: integer, minimum: 0 }
      example: { name: BinSpace, chineseName: 玢的空间, description: 记录代码，也记录生活。, version: 3 }
    UpdateHeroDTO:
      type: object
      required: [eyebrow, heroTitle, heroSubtitle, desktopVideo, mobileVideo, poster, version]
      properties:
        eyebrow: { type: string, nullable: true, maxLength: 128 }
        heroTitle: { type: string, minLength: 1, maxLength: 180 }
        heroSubtitle: { type: string, nullable: true, maxLength: 500 }
        desktopVideo: { type: string, nullable: true, maxLength: 1024 }
        mobileVideo: { type: string, nullable: true, maxLength: 1024 }
        poster: { type: string, nullable: true, maxLength: 1024 }
        version: { type: integer, minimum: 0 }
      example:
        eyebrow: BINSPACE · PERSONAL DAYBOOK
        heroTitle: 在暖光里，留下代码与生活。
        heroSubtitle: 记录认真写下的技术文章和日常时刻。
        desktopVideo: /media/binspace-hero.mp4
        mobileVideo: /media/binspace-hero-mobile.mp4
        poster: /media/binspace-hero.jpg
        version: 3
    UpdateAnnouncementDTO:
      type: object
      required: [content, enabled, version]
      properties:
        content: { type: string, nullable: true, maxLength: 2000 }
        enabled: { type: boolean }
        version: { type: integer, minimum: 0 }
      example: { content: 欢迎来到玢的空间。, enabled: true, version: 4 }
    UpdateMusicDTO:
      type: object
      required: [title, artist, audioUrl, cover, version]
      properties:
        title: { type: string, nullable: true, maxLength: 180 }
        artist: { type: string, nullable: true, maxLength: 180 }
        audioUrl: { type: string, nullable: true, maxLength: 1024 }
        cover: { type: string, nullable: true, maxLength: 1024 }
        version: { type: integer, minimum: 0 }
      example: { title: 空间背景音乐, artist: Artist, audioUrl: /media/music/theme.mp3, cover: /media/music/theme.jpg, version: 5 }
    UpdateAppearanceDTO:
      allOf:
        - { $ref: '#/components/schemas/AppearanceVO' }
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 0, example: 6 }
    UpdatePageMediaDTO:
      allOf:
        - { $ref: '#/components/schemas/PageMediaVO' }
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 0, example: 7 }
    SiteConfigSectionBase:
      type: object
      required: [version, updatedAt]
      properties:
        version: { type: integer, minimum: 0, example: 4 }
        updatedAt: { type: string, format: date-time, example: '2026-09-02T10:20:30.123Z' }
    SiteBasicSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config: { $ref: '#/components/schemas/SiteBasicVO' }
    HeroSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config: { $ref: '#/components/schemas/HeroVO' }
    AnnouncementSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config:
              anyOf:
                - { $ref: '#/components/schemas/AnnouncementVO' }
                - { $ref: '#/components/schemas/NullOnly' }
    MusicSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config:
              anyOf:
                - { $ref: '#/components/schemas/MusicVO' }
                - { $ref: '#/components/schemas/NullOnly' }
    AppearanceSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config: { $ref: '#/components/schemas/AppearanceVO' }
    PageMediaSectionVO:
      allOf:
        - { $ref: '#/components/schemas/SiteConfigSectionBase' }
        - type: object
          required: [config]
          properties:
            config: { $ref: '#/components/schemas/PageMediaVO' }
...
    SiteBasicSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/SiteBasicSectionVO' } }
    HeroSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/HeroSectionVO' } }
    AnnouncementSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/AnnouncementSectionVO' } }
    MusicSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/MusicSectionVO' } }
    AppearanceSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/AppearanceSectionVO' } }
    PageMediaSectionResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/PageMediaSectionVO' } }
...
    TagVO:
      type: object
      required: [id, name, slug, active]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: Lua }
        slug: { type: string, example: lua }
        active: { type: boolean, example: true }
...
    TagAdminVO:
      allOf:
        - { $ref: '#/components/schemas/TagVO' }
        - type: object
          required: [updatedAt]
          properties:
            updatedAt: { type: string, format: date-time }
...
    CreateTagDTO:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, nullable: true, maxLength: 80 }
      example: { name: Lua, slug: lua }
    UpdateTagDTO:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, minLength: 1, maxLength: 80 }
      example: { name: Lua, slug: lua }
...
    TagPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TagVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
    TagAdminPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TagAdminVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
...
    TagPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/TagPageVO' } }
    TagAdminPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/TagAdminPageVO' } }
    TagAdminResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/TagAdminVO' } }
...